refactor: reorganize notification-related classes and update package structure
