--- src/main/java/app/bottlenote/user/dto/request/SingleTokenRequest.java
@@ -1,6 +1,0 @@
-package app.bottlenote.user.dto.request;
-
-public record SingleTokenRequest(
-	String token
-) {
-}
