--- src/main/java/app/bottlenote/user/dto/response/TokenSaveResponse.java
@@ -1,18 +1,0 @@
-package app.bottlenote.user.dto.response;
-
-import app.bottlenote.user_notification.domain.constant.Platform;
-import app.external.push.dto.model.TokenMessage;
-
-public record TokenSaveResponse(
-	String deviceToken,
-	Platform platform,
-	TokenMessage message
-) {
-	public static TokenSaveResponse of(
-		String deviceToken,
-		Platform platform,
-		TokenMessage message
-	) {
-		return new TokenSaveResponse(deviceToken, platform, message);
-	}
-}
