--- src/main/java/app/bottlenote/user/service/NotificationService.java
@@ -1,7 +1,0 @@
-package app.bottlenote.user.service;
-
-import app.bottlenote.user.dto.request.NotificationMessage;
-
-public interface NotificationService {
-	void sendNotification(NotificationMessage message);
-}
