+++ src/main/java/app/bottlenote/user/service/UserFacade.java
@@ -2,8 +2,6 @@ package app.bottlenote.user.service;
 
 import app.bottlenote.user.dto.response.UserProfileInfo;
 
-import java.util.List;
-
 /**
  * 유저 도메인 서포트를 위한 인터페이스입니다.
  * 이 인터페이스는 유저 식별자를 기반으로 유저 데이터를 검증하는 메소드를 제공합니다.
@@ -28,7 +26,4 @@ public interface UserFacade {
 	void isValidUserId(Long userId);
 
 	UserProfileInfo getUserProfileInfo(Long userId);
-
-	//List<String> 발신 가능한 유저 토큰 목록 조회
-	List<String> getAvailableUserTokens(List<Long> userId);
 }
