--- src/main/java/app/bottlenote/user_notification/domain/NotificationRepository.java
@@ -1,13 +1,0 @@
-package app.bottlenote.user_notification.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface NotificationRepository {
-
-	Notification save(Notification notification);
-
-	Optional<Notification> findById(Long notifyId);
-
-	List<Notification> findAll();
-}
