--- src/main/java/app/bottlenote/user_notification/domain/constant/NotificationCategory.java
@@ -1,17 +1,0 @@
-package app.bottlenote.user_notification.domain.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum NotificationCategory {
-	REVIEW("리뷰"),
-	NOTICE("공지사항"),
-	QUESTION("문의"),
-	ANSWER("답변");
-
-	private final String description;
-
-	NotificationCategory(String description) {
-		this.description = description;
-	}
-}
