+++ src/main/java/app/external/notification/application/NotificationService.java
@@ -1,0 +1,7 @@
+package app.external.notification.application;
+
+import app.external.notification.data.payload.NotificationMessage;
+
+public interface NotificationService {
+	void sendNotification(NotificationMessage message);
+}
