+++ src/main/java/app/external/notification/domain/constant/NotificationStatus.java
@@ -1,0 +1,18 @@
+package app.external.notification.domain.constant;
+
+import lombok.Getter;
+
+@Getter
+public enum NotificationStatus {
+	PENDING("대기 중"),
+	SENT("전송됨"),
+	READ("읽음"),
+	FAILED("실패");
+
+	private final String description;
+
+	NotificationStatus(String description) {
+		this.description = description;
+	}
+
+}
