+++ src/main/java/app/external/notification/domain/constant/Platform.java
@@ -1,0 +1,15 @@
+package app.external.notification.domain.constant;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+public enum Platform {
+	ANDROID, IOS;
+
+	@JsonCreator
+	public static Platform fromString(String value) {
+		if (value == null) {
+			return null;
+		}
+		return Platform.valueOf(value.toUpperCase());
+	}
+}
