+++ src/main/java/app/external/notification/presentation/NotificationController.java
@@ -1,0 +1,12 @@
+package app.external.notification.presentation;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/external/notification")
+@RequiredArgsConstructor
+public class NotificationController {
+
+}
