+++ src/main/java/app/external/notification/repository/NotificationRepository.java
@@ -1,0 +1,9 @@
+package app.external.notification.repository;
+
+import app.external.notification.domain.Notification;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
+}
