+++ src/main/java/app/external/push/application/PushHandler.java
@@ -1,0 +1,10 @@
+package app.external.push.application;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface PushHandler {
+
+	void sendPush(List<Long> userIds, String message);
+	void schedulePush(List<Long> userIds, String message, LocalDateTime scheduledTime);
+}
