+++ src/main/java/app/external/push/data/request/SingleTokenRequest.java
@@ -1,0 +1,6 @@
+package app.external.push.data.request;
+
+public record SingleTokenRequest(
+	String token
+) {
+}
