--- src/main/java/app/external/push/service/PushHandler.java
@@ -1,10 +1,0 @@
-package app.external.push.service;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface PushHandler {
-
-	void sendPush(List<Long> userIds, String message);
-	void schedulePush(List<Long> userIds, String message, LocalDateTime scheduledTime);
-}
